More Performance Optimizations and Code Cleaing! Next Up: Hashed Grid!

+++ Senku2D/Senku2D/Rigid Body/Collision Processing/Potential Contact List.cpp
@@ -1,6 +1,6 @@
 #include "Potential Contact List.h"
 
-Senku2D::PotentialContactList::PotentialContactList(const uint8_t& Limit) :
+Senku2D::PotentialContactList::PotentialContactList(const size_t& Limit) :
 	m_Limit(Limit),
 	m_InternalList()
 {
@@ -14,7 +14,7 @@ Senku2D::PotentialContactList::~PotentialContactList()
 	m_InternalList.clear();
 }
 
-Senku2D::PotentialRigidBodyContact& Senku2D::PotentialContactList::GetContact(const uint8_t& Index)
+Senku2D::PotentialRigidBodyContact& Senku2D::PotentialContactList::GetContact(const size_t& Index)
 {
 	//Index Should Not Exceed Limit
 	assert(Index < m_Limit);
@@ -30,7 +30,7 @@ void Senku2D::PotentialContactList::Clear()
 	}
 }
 
-const uint8_t Senku2D::PotentialContactList::GetLimit() const
+const size_t Senku2D::PotentialContactList::GetLimit() const
 {
 	return m_Limit;
 }
